requirements.txt
setuptools
wheel
requests
lxml
cssselect
telepot
python-dotenv
pickledb